Add necessary compilation of gschema files
Without compiling the gchema.xml file the new settings don't get picked up by the system. I tried several times to get this example (and my own program) working, but with no luck. Googling led me to the concept of compiling the gschema.xml files. When I looked at the code of a few elementary projects I found the described solution. This is what got things working for me. So I think it's pretty important to explain it here in the tutorial.