Update README.md
In the previous page, it mentions that you should add a GPL header to your files. For consistency, the metadata should display that license (also, as Lessig once told me, CC0 – or any Creative Commons licenses – should not be used for software as they do not include limitation of liability clauses).